refactor: use appropriate types for scales and dimensions
Change scale factors from f32 to f64 and dimensions from f32 to u32 in
`TargetPosition` and `WindowState` structs to match their actual types
throughout the codebase and eliminate unnecessary casting.

- Scale factors from Bevy monitors are f64; store them directly without
  casting to f32 then back to f64 during calculations
- Window dimensions are u32 in Bevy's Window API; store them as u32 in
  saved state rather than converting through f32
- Use `f64::from()` for explicit integer-to-f64 conversions in scale
  compensation calculations for clarity
- Remove redundant cast-related clippy exceptions that are no longer
  needed with proper type usage
- Simplifies type conversions and reduces precision loss from intermediate
  f32 casts

+++ src/types.rs
@@ -81,9 +81,9 @@ pub struct TargetPosition {
     /// Window entity being restored.
     pub entity:                 Entity,
     /// Scale factor of the target monitor.
-    pub target_scale:           f32,
+    pub target_scale:           f64,
     /// Scale factor of the monitor where the window starts (keyboard focus monitor).
-    pub starting_scale:         f32,
+    pub starting_scale:         f64,
     /// Strategy for handling scale factor differences between monitors.
     pub monitor_scale_strategy: MonitorScaleStrategy,
 }
@@ -99,7 +99,7 @@ pub struct RestoreWindowConfig {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct WindowState {
     pub position:      Option<(i32, i32)>,
-    pub width:         f32,
-    pub height:        f32,
+    pub width:         u32,
+    pub height:        u32,
     pub monitor_index: usize,
 }
